Move CPathFind crash fix hook into CMultiplayerSA_CrashFixHacks
Moved per FileEX's review comment on PR #5245.
