--- Client/game_sa/CPathFindSA.h
@@ -1,18 +1,0 @@
-/*****************************************************************************
- *
- *  PROJECT:     Multi Theft Auto v1.0
- *  LICENSE:     See LICENSE in the top level directory
- *  FILE:        game_sa/CPathFindSA.h
- *  PURPOSE:     Header file for path find (CPathFind) hooks
- *
- *  Multi Theft Auto is available from https://www.multitheftauto.com/
- *
- *****************************************************************************/
-
-#pragma once
-
-class CPathFindSA
-{
-public:
-    static void StaticSetHooks();
-};
