+++ Shared/sdk/SharedUtil.MemAccess.hpp
@@ -50,7 +50,7 @@ namespace SharedUtil
         VirtualProtect((LPVOID)0x729000, 0x1000, PAGE_EXECUTE_READWRITE, &oldProt);
         VirtualProtect((LPVOID)0x742000, 0x1000, PAGE_EXECUTE_READWRITE, &oldProt);
 
-        // Used by CPathFind's hooks in game_sa/CPathFindSA.cpp, installed via the Fast path.
+        // Used by the CPathFind crash fix in multiplayer_sa/CMultiplayerSA_CrashFixHacks.cpp, installed via the Fast path.
         VirtualProtect((LPVOID)0x156F000, 0x1000, PAGE_EXECUTE_READWRITE, &oldProt);
     }
